get rid of abstact class AngleBaseWaveGenerator and make use of contract directly
