+++ zentone/src/main/java/com/github/nisrulz/zentone/wavegenerators/SquareWaveGenerator.kt
@@ -11,7 +11,10 @@ import kotlin.math.sin
  * @see <a
  *     href="https://en.wikipedia.org/wiki/Square_wave">Wikipedia</a>
  */
-object SquareWaveGenerator : AngleBaseWaveGenerator() {
+object SquareWaveGenerator : WaveByteArrayGenerator {
+
+    override var angle: Double = 0.0
+    override var angleStep: Double = 0.0
 
     override fun calculateData(angle: Double, amplitude: Int): Byte {
         return (amplitude * waveFunction(angle) * Byte.MAX_VALUE).toInt().toByte()
