+++ zentone/src/main/java/com/github/nisrulz/zentone/wavegenerators/WaveByteArrayGenerator.kt
@@ -4,8 +4,12 @@ import com.github.nisrulz.zentone.DEFAULT_AMPLITUDE
 import com.github.nisrulz.zentone.DEFAULT_FREQUENCY_HZ
 import com.github.nisrulz.zentone.DEFAULT_SAMPLE_RATE
 import com.github.nisrulz.zentone.internal.minBufferSize
+import kotlin.math.PI
 
 interface WaveByteArrayGenerator {
+
+    var angle: Double
+    var angleStep: Double
 
     /**
      * Generate byte data for tone
@@ -24,16 +28,40 @@ interface WaveByteArrayGenerator {
         val generatedSnd = ByteArray(bufferSize)
 
         generatedSnd.indices.forEach { i ->
-            generatedSnd[i] = calculateData(i, DEFAULT_AMPLITUDE)
+            generatedSnd[i] = calculateData(DEFAULT_AMPLITUDE)
         }
 
         return generatedSnd
     }
 
-    fun calculateData(index: Int, amplitude: Int): Byte
+    fun calculateData(angle: Double, amplitude: Int): Byte
 
-    fun setup(freqOfTone: Float, sampleRate: Int)
 
-    fun reset()
+    /**
+     * Setup required before generating a frame. Must be called at least once before calculateData.
+     */
+    fun setup(freqOfTone: Float, sampleRate: Int) {
+        val samplingInterval = sampleRate / freqOfTone
+        angleStep = PI / samplingInterval
+    }
+
+    /**
+     * Reset the generator to be able to start over from the start again.
+     */
+    fun reset() {
+        angle = 0.0
+    }
+
+    private fun incrementAngle(
+        angle: Double,
+        angleStep: Double
+    ): Double {
+        return (angle + angleStep) % (2 * Math.PI)
+    }
+
+    private fun calculateData(amplitude: Int): Byte {
+        angle = incrementAngle(angle, angleStep)
+        return calculateData(angle, amplitude)
+    }
 
 }
